backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg[binary]
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
httpx
sse-starlette
alembic
slowapi
email-validator
